refactor(svdsbtl): move write_bytes_atomic to CPfilepaths (free function)
Per review: the atomic write-temp + rename helper is a generic file
utility, not an SVDSBTL serializer responsibility.  Other callers
that want atomic cache writes (BICUBIC/TTSE tables, future cache
files) shouldn't have to depend on the SBTL serializer header.

Moved:
  - declaration: include/CPfilepaths.h (alongside get_binary_file_contents)
  - definition + anon-namespace process-salt helpers: src/CPfilepaths.cpp
  - now a free function in the global namespace (matches the rest of
    CPfilepaths.h's style — no namespace, all free functions)

Call sites updated to bare `::write_bytes_atomic(...)`:
  - src/SBTL/SVDSurfaceSerializer.cpp::save_to_file
  - src/Backends/SVDSBTL/SVDSBTLBackend.cpp::save_critpatch_cache_

Header impact:
  - include/CoolProp/sbtl/SVDSurfaceSerializer.h no longer needs
    <filesystem> or <cstddef>; pulled the includes back to the
    pre-fix set.
  - src/SBTL/SVDSurfaceSerializer.cpp dropped <atomic>, <random>,
    <sstream> (helper internals migrated to CPfilepaths.cpp).
  - src/Backends/SVDSBTL/SVDSBTLBackend.cpp gains #include "CPfilepaths.h".

Test: src/Tests/CoolProp-Tests-SVDSBTL.cpp updated to call the bare
free function (was qualified `cp_sbtl::SVDSurfaceSerializer::...`).
Race test still passes deterministically.
